Documents and splits up Chat.cs into ChatProvider.cs
